requirements.txt
Django==4.0.4
django-celery-results==2.4.0
celery==5.2.7
redis == 4.3.4
asgiref==3.5.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
python-decouple==3.6
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0